requirements.txt
-i https://oneartifactoryprod.verizon.com/artifactory/api/pypi/Pypi-remote/simple
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.1
Flask-Cors==3.0.10
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
requests==2.29.0
six==1.16.0
urllib3==1.26.15
Werkzeug==2.3.2
